+++ openchrom/plugins/net.openchrom.xxd.process.supplier.templates.ui/src/net/openchrom/xxd/process/supplier/templates/ui/internal/provider/PeakIntegratorComparator.java
@@ -30,10 +30,10 @@ public class PeakIntegratorComparator extends AbstractRecordTableComparator impl
 			//
 			switch(getPropertyIndex()) {
 				case 0:
-					sortOrder = Double.compare(setting2.getStartRetentionTime(), setting1.getStartRetentionTime());
+					sortOrder = Integer.compare(setting2.getStartRetentionTime(), setting1.getStartRetentionTime());
 					break;
 				case 1:
-					sortOrder = Double.compare(setting2.getStopRetentionTime(), setting1.getStopRetentionTime());
+					sortOrder = Integer.compare(setting2.getStopRetentionTime(), setting1.getStopRetentionTime());
 					break;
 				case 2:
 					sortOrder = setting2.getIdentifier().compareTo(setting1.getIdentifier());
