+++ openchrom/plugins/net.openchrom.xxd.process.supplier.templates.ui/src/net/openchrom/xxd/process/supplier/templates/ui/internal/provider/StandardsAssignerComparator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 Lablicate GmbH.
+ * Copyright (c) 2018, 2019 Lablicate GmbH.
  * 
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -33,10 +33,10 @@ public class StandardsAssignerComparator extends AbstractRecordTableComparator i
 					sortOrder = setting2.getName().compareTo(setting1.getName());
 					break;
 				case 1:
-					sortOrder = Double.compare(setting2.getStartRetentionTime(), setting1.getStartRetentionTime());
+					sortOrder = Integer.compare(setting2.getStartRetentionTime(), setting1.getStartRetentionTime());
 					break;
 				case 2:
-					sortOrder = Double.compare(setting2.getStopRetentionTime(), setting1.getStopRetentionTime());
+					sortOrder = Integer.compare(setting2.getStopRetentionTime(), setting1.getStopRetentionTime());
 					break;
 				case 3:
 					sortOrder = Double.compare(setting2.getConcentration(), setting1.getConcentration());
