+++ openchrom/plugins/net.openchrom.xxd.process.supplier.templates/src/net/openchrom/xxd/process/supplier/templates/comparator/ReferenceComparator.java
@@ -20,7 +20,7 @@ public class ReferenceComparator implements Comparator<AssignerReference> {
 	@Override
 	public int compare(AssignerReference setting1, AssignerReference setting2) {
 
-		int result = Double.compare(setting1.getStartRetentionTime(), setting2.getStartRetentionTime());
+		int result = Integer.compare(setting1.getStartRetentionTime(), setting2.getStartRetentionTime());
 		if(result == 0) {
 			result = setting1.getInternalStandard().compareTo(setting2.getInternalStandard());
 		}
